Split the camera and program state functions into seperate files. solar.cpp should be pretty much done with now except for actually creating the planets.

+++ Solar/Solar/Structs.h
@@ -5,8 +5,8 @@ struct ProgramState
 {
 	bool Wireframe, Flat, Texture, Paused;
 	double Resolution;
-	int AnimTime;
-	long long Frame;
+	unsigned int AnimTime;
+	int Frame;
 
 	ProgramState()
 	{
